% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlineSpecs.R
\name{inlineSpecs}
\alias{inlineSpecs}
\title{Add together two numbers}
\usage{
inlineSpecs(data, output_type = "kable")
}
\arguments{
\item{data}{The dataframe}

\item{output_type}{Output type required. For now, if not kable you get a dataframe}
}
\value{
Summarise a dataframe in a table
}
\description{
Add together two numbers
}
\examples{
d_data <- data.frame(
  Time = as.Date('2009-01-01') + 0:9,
  Colours = rep(c("red", "blue"),5),
  Numbers = rnorm(10, 0, 10)
)
d_data$Potatoes <- rep(c("sweet",NA,"russet","french-fried","roasted"),2)
inlineSpecs(d_data)
}
