% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryspecs.R
\name{summaryspecs}
\alias{summaryspecs}
\title{Add together two numbers}
\usage{
summaryspecs(data, output_type = "kable")
}
\arguments{
\item{data}{The dataframe}

\item{output_type}{Output type required. For now, if not kable you get a dataframe}
}
\value{
Summarise a dataframe in a table
}
\description{
Add together two numbers
}
